+++ Site/webapp/wdkCustomization/js/client/components/Visualizations/Table/ColumnAccessors/index.ts
@@ -24,4 +24,5 @@ export interface ColumnAccessor {
     htmlColor?: string;
     className?: string;
     maxLength?: number;
+
 }
