+++ Site/webapp/wdkCustomization/js/client/components/Visualizations/Table/RowSelectors/RowSelectCheckbox.tsx
