+++ Site/webapp/wdkCustomization/js/client/components/Visualizations/Table/RowSelectors/index.ts
@@ -1,0 +1,2 @@
+export * from "./RowSelectCheckbox";
+export * from "./RowSelectButton";
